Backend/requirements.txt
bcrypt==4.2.0
blinker==1.8.2
boto3==1.35.28
botocore==1.35.28
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Mail==0.10.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
PyJWT==2.9.0
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.3
Werkzeug==3.0.4
